backend/requirements.txt
asgiref==3.8.1
borb==2.1.24
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
CodeConvert==3.0.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==3.2
django-admin==2.0.2
django-cors-headers==3.13.0
django-excel-response2==3.0.6
django-filter==21.1
django-models-ext==1.1.11
django-short-url==1.1.8
django-six==1.0.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
excel-base==1.0.4
flake8==6.0.0
flake8-isort==6.0.0
fonttools==4.53.0
furl==2.1.3
gunicorn==20.1.0
idna==3.7
isort==5.13.2
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
mccabe==0.7.0
oauthlib==3.2.2
orderedmultidict==1.0.1
pillow==10.3.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.8.0
pypng==0.20220715.0
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
screen==1.0.1
short_url==1.2.2
shortuuid==1.0.13
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.4
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
xlwt==1.3.0
